requirements.txt
lark==1.1.2
pandas==1.3.5
PyQt6==6.3.1
PyYAML==6.0
